requirements.txt
# REQUIRED PIP INSTALLATIONS

# Enter the command in your termial in this project folder
# python -m venv wa-viewer-env
# pip install -r requirements.txt

art>=5.9
ipykernel
pyautogui
pytz>=2021.3
python-dotenv>=1.1.1
selenium>=4.8.2
beautifulsoup4>=4.13.5
git+https://github.com/KrAsH-CoD3/gabrielrih-callmebot.git@refactor/codebase